@@ -8,46 +8,37 @@ use rustyline::hint::Hinter;
 use rustyline::validate::Validator;
 use rustyline::{Context, Helper, Editor, history::DefaultHistory};
 
-const VERSION: &str = "3.1.0";
+const VERSION: &str = "3.1.1";
 
 fn main() {
-    let args: Vec<String> = std::env::args().collect();
-    
-    if args.len() == 2 && args[1] == "--version" {
+    if std::env::args().nth(1).as_deref() == Some("--version") {
         display_version();
         return;
     }
-    
-    let packages = load_packages();
+
     let mut rl: Editor<PackageCompleter, DefaultHistory> = Editor::new().expect("Failed to create editor");
-    rl.set_helper(Some(PackageCompleter { packages }));
-    
-    println!("\nWelcome to Archie v{}", VERSION);
-    println!("Using paru package manager");
-    println!("Type 'h' for help\n");
-    
+    rl.set_helper(Some(PackageCompleter { packages: load_packages() }));
+
+    println!("\nWelcome to Archie v{VERSION}\nType 'h' for help\n");
+
     loop {
-        let input: String = match rl.readline("$ ") {
+        let input = match rl.readline("$ ") {
             Ok(line) => line,
-            Err(ReadlineError::Interrupted) => break,
-            Err(ReadlineError::Eof) => break,
-            Err(err) => {
-                eprintln!("Error: {:?}", err);
-                break;
-            }
+            Err(ReadlineError::Interrupted | ReadlineError::Eof) => break,
+            Err(e) => { eprintln!("Error: {e:?}"); break; }
         };
-        
+
         rl.add_history_entry(&input).expect("Failed to add history");
-        
+
         match input.trim() {
-            "u" => exec(&["-Syu"]),
-            "i" => prompt_install(&mut rl),
-            "r" => prompt_remove(&mut rl, "-R"),
-            "p" => prompt_remove(&mut rl, "-Rns"),
-            "s" => prompt_search(&mut rl),
-            "c" => exec(&["-Sc"]),
-            "o" => exec_shell("paru -Rns $(pacman -Qtdq)"),
-            "h" => show_help(),
+            "u" => paru(&["-Syu"]),
+            "i" => prompt(&mut rl, "Package: ", |p| paru(&["-S", p])),
+            "r" => prompt(&mut rl, "Package: ", |p| paru(&["-R", p])),
+            "p" => prompt(&mut rl, "Package: ", |p| paru(&["-Rns", p])),
+            "s" => prompt(&mut rl, "Search: ",  |p| paru(&["-Ss", p])),
+            "c" => paru(&["-Sc"]),
+            "o" => shell("paru -Rns $(pacman -Qtdq)"),
+            "h" => println!("\nCommands:\n  u - Update system      i - Install package\n  r - Remove package     p - Purge package\n  s - Search packages    c - Clean cache\n  o - Remove orphans     h - Show help\n  q - Quit\n"),
             "q" => break,
             "" => continue,
             _ => println!("Unknown command. Type 'h' for help"),
@@ -57,105 +48,55 @@ fn main() {
 }
 
 fn display_version() {
+    let paru = Command::new("paru").arg("--version").output()
+        .ok()
+        .and_then(|o| String::from_utf8(o.stdout).ok())
+        .and_then(|s| s.lines().next().map(|l| l.split_whitespace().skip(1).collect::<Vec<_>>().join(" ")))
+        .unwrap_or_else(|| "unknown".into());
+
     println!("    __     ");
-    println!(" .:--.'.   Archie-ng v{} - Fast & Easy package management for Arch Linux", VERSION);
+    println!(" .:--.'.   Archie-ng v{VERSION} - Fast & Easy package management for Arch Linux");
     println!("/ |   \\ |  Written in Rust, powered by paru.");
-    println!("`\" __ | |  {}", get_paru_version());
+    println!("`\" __ | |  {paru}");
     println!(" .'.''| |  ");
     println!("/ /   | |_ This program may be freely redistributed under the terms of the GNU General Public License.");
     println!("\\ \\._,\\ '/ Created & maintained by Gurov");
     println!(" `--'  `\"  ");
 }
 
-fn get_paru_version() -> String {
-    if let Ok(output) = Command::new("paru").arg("--version").output() {
-        if let Ok(line) = String::from_utf8(output.stdout) {
-            if let Some(version) = line.lines().next() {
-                return version.split_whitespace().skip(1).collect::<Vec<_>>().join(" ");
-            }
-        }
-    }
-    String::from("unknown")
-}
-
 fn load_packages() -> Vec<String> {
-    let mut packages = Vec::new();
-    
-    if let Ok(output) = Command::new("pacman").arg("-Slq").output() {
-        if let Ok(content) = String::from_utf8(output.stdout) {
-            for line in content.lines() {
-                if !line.is_empty() {
-                    packages.push(line.to_string());
-                }
-            }
-        }
-    }
-    
+    let mut packages: Vec<String> = Command::new("pacman").arg("-Slq").output()
+        .ok()
+        .and_then(|o| String::from_utf8(o.stdout).ok())
+        .map(|s| s.lines().filter(|l| !l.is_empty()).map(String::from).collect())
+        .unwrap_or_default();
+
     let aur_cache = PathBuf::from(env!("HOME")).join(".cache/paru/packages.aur");
-    if let Ok(content) = fs::read_to_string(&aur_cache) {
-        for line in content.lines() {
-            if !line.is_empty() {
-                packages.push(line.to_string());
-            }
-        }
+    if let Ok(content) = fs::read_to_string(aur_cache) {
+        packages.extend(content.lines().filter(|l| !l.is_empty()).map(String::from));
     }
-    
+
     packages.sort();
     packages.dedup();
     packages
 }
 
-fn exec(args: &[&str]) {
+fn paru(args: &[&str]) {
     let _ = Command::new("paru").args(args).status();
 }
 
-fn exec_shell(cmd: &str) {
+fn shell(cmd: &str) {
     let _ = Command::new("sh").arg("-c").arg(cmd).status();
 }
 
-fn prompt_install(rl: &mut Editor<PackageCompleter, DefaultHistory>) {
-    let pkg: String = match rl.readline("Package: ") {
-        Ok(line) => line,
-        Err(_) => return,
-    };
-    rl.add_history_entry(&pkg).expect("Failed to add history");
-    let pkg = pkg.trim();
-    if !pkg.is_empty() {
-        exec(&["-S", pkg]);
+fn prompt(rl: &mut Editor<PackageCompleter, DefaultHistory>, label: &str, action: impl Fn(&str)) {
+    if let Ok(input) = rl.readline(label) {
+        rl.add_history_entry(&input).expect("Failed to add history");
+        let trimmed = input.trim();
+        if !trimmed.is_empty() {
+            action(trimmed);
+        }
     }
-}
-
-fn prompt_remove(rl: &mut Editor<PackageCompleter, DefaultHistory>, flag: &str) {
-    let pkg: String = match rl.readline("Package: ") {
-        Ok(line) => line,
-        Err(_) => return,
-    };
-    rl.add_history_entry(&pkg).expect("Failed to add history");
-    let pkg = pkg.trim();
-    if !pkg.is_empty() {
-        exec(&[flag, pkg]);
-    }
-}
-
-fn prompt_search(rl: &mut Editor<PackageCompleter, DefaultHistory>) {
-    let query: String = match rl.readline("Search: ") {
-        Ok(line) => line,
-        Err(_) => return,
-    };
-    rl.add_history_entry(&query).expect("Failed to add history");
-    let query = query.trim();
-    if !query.is_empty() {
-        exec(&["-Ss", query]);
-    }
-}
-
-fn show_help() {
-    println!("\nCommands:");
-    println!("  u - Update system      i - Install package");
-    println!("  r - Remove package     p - Purge package");
-    println!("  s - Search packages    c - Clean cache");
-    println!("  o - Remove orphans     h - Show help");
-    println!("  q - Quit\n");
 }
 
 struct PackageCompleter {
@@ -163,24 +104,14 @@ struct PackageCompleter {
 }
 
 impl Helper for PackageCompleter {}
+impl Hinter for PackageCompleter { type Hint = String; }
+impl Validator for PackageCompleter {}
+impl Highlighter for PackageCompleter {}
 
 impl Completer for PackageCompleter {
     type Candidate = String;
-    
+
     fn complete(&self, line: &str, _pos: usize, _ctx: &Context<'_>) -> Result<(usize, Vec<String>), ReadlineError> {
-        let matches: Vec<String> = self.packages
-            .iter()
-            .filter(|pkg| pkg.starts_with(line))
-            .cloned()
-            .collect();
-        Ok((0, matches))
+        Ok((0, self.packages.iter().filter(|p| p.starts_with(line)).cloned().collect()))
     }
 }
-
-impl Hinter for PackageCompleter {
-    type Hint = String;
-}
-
-impl Validator for PackageCompleter {}
-
-impl Highlighter for PackageCompleter {}
